Split handler.py into several
