--- modulestf/terraform.py
@@ -1,1 +1,0 @@
-x = 1
