+++ templates/terragrunt-single-layer/cookiecutter.json
@@ -2,5 +2,6 @@
   "dir_name": "template",
   "module_source": "",
   "module_variables": {},
-  "params": {}
+  "params": {},
+  "dependencies": {}
 }
